refactor(web): extract Chip + ChipPicker + filter-helpers from advanced-filter-panel (#692)
`advanced-filter-panel.tsx` (623 LOC, the largest non-page component
in the web codebase) bundled three things that should live elsewhere:

1. **Chip + ChipPicker** — the multi-select chip toggle pattern
   (defined inline as a private component) plus the toggle button
   styling that was duplicated in the play-status row. Extracted
   to `@/components/ui/chip.tsx`:
   - `Chip` — pill-shaped toggle button (single source of truth for
     the selected/unselected visual states)
   - `ChipPicker` — multi-select wrapper with the "Show all (N)"
     overflow toggle, built on top of `Chip`
   The play-status row now uses `Chip` directly instead of
   reimplementing the same button + class set.

2. **Filter helpers** — `filtersToRecord`, `savedSearchToFilters`,
   `countActiveFilters`, plus the static `REGION_OPTIONS` /
   `GENRE_OPTIONS` / `PLAY_STATUS_OPTIONS` tables. Moved to
   `features/games/lib/filter-helpers.ts`. `savedSearchToFilters` is
   re-exported from the panel module so existing call sites
   (games-page, console-games-page, the panel test) keep working
   without changes.

3. **RangeInput's raw `<input type=\"number\">`** — replaced with the
   shared `Input` component from `@/components/ui`. Drops the
   hand-rolled focus-ring + border Tailwind that was duplicating
   what `Input` already provides.

Numbers:
- advanced-filter-panel.tsx: 623 → 471 LOC (-152)
- New: chip.tsx (116), filter-helpers.ts (95)
- Net codebase delta: +59 LOC, but the extracted pieces are now
  reusable across the app

Behaviour preserved. All 1557 web tests pass (no test edits
needed). `npx tsc --noEmit` clean.

Closes #692
